% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_date_named.R
\name{assert_date_named}
\alias{assert_date_named}
\title{Assert Date Value in Column}
\usage{
assert_date_named(column, df, prefix_column = NULL, ...)
}
\arguments{
\item{column}{A character vector or string with the column name to be tested.}

\item{df}{The data frame that contains the column.}

\item{prefix_column}{A character string that will be prepended to the column name in the assertion message. Default is NULL.}

\item{...}{Additional parameters are passed to the \code{checkmate::assert_date} function.}
}
\value{
None
}
\description{
This function asserts that the values in a specified column of a data frame are of Date type.
It uses the \code{checkmate::assert_date} function to perform the assertion.
}
