% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_no_duplicates_in_group.R
\name{assert_no_duplicates_in_group}
\alias{assert_no_duplicates_in_group}
\title{Assert No Duplicates in Group}
\usage{
assert_no_duplicates_in_group(df, group_vars, assertion_fail = "stop")
}
\arguments{
\item{df}{A data frame.}

\item{group_vars}{A character vector of column names.}

\item{assertion_fail}{A character string indicating the action to take if the assertion fails. Can be "stop" (default) or "warn".}
}
\value{
The input data frame.
}
\description{
This function asserts that there are no duplicate rows in the specified columns of a data frame.
It groups the data frame by the specified columns, counts the number of unique values for each group, and checks if there are any groups with more than one row.
If there are, it prints an error message and stops the execution (unless \code{assertion_fail} is set to "warn").
}
