% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_type_consistency.R
\name{assert_type_consistency}
\alias{assert_type_consistency}
\title{Assert Type Consistency Between Data and Metadata}
\usage{
assert_type_consistency(data, metadata)
}
\arguments{
\item{data}{A data frame. The dataset containing the fields to validate.}

\item{metadata}{A data frame. Metadata specifying the expected data types for each field.
It must include the columns \code{raw_field_name}, \code{type_of_variable}, and \code{preferred_field_name}.}
}
\value{
No return value. The function issues warnings if any fields have an unexpected type.
}
\description{
This function checks whether the data types of fields in a dataset match the expected types specified in the metadata.
It warns if any fields have a different type than expected.
}
