% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_message.R
\name{assertion_message}
\alias{assertion_message}
\title{Assert Message Based on Type}
\usage{
assertion_message(message, assertion_fail = "stop")
}
\arguments{
\item{message}{A character string representing the message to be asserted.}

\item{assertion_fail}{A character string indicating the action to take if the assertion fails. Can be an AssertCollection, "warning", or "stop" (default).}
}
\value{
None
}
\description{
This function asserts a message based on the type specified.
It can either push the message to an AssertCollection, print a warning, or stop execution with an error message.
}
