% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_classes.R
\name{check_classes}
\alias{check_classes}
\title{Check classes}
\usage{
check_classes(x, naming)
}
\arguments{
\item{x}{Data frame.}

\item{naming}{dataframecontaining the column "Type_informatie".}
}
\value{
Either nothing or a warning when the classes do not correspond.
}
\description{
Checks whether classes correspond to those in the naming documentation file.
}
