% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_documentation_analysisset.R
\name{check_documentation_analysisset}
\alias{check_documentation_analysisset}
\title{Check documentation Analysis set}
\usage{
check_documentation_analysisset(x)
}
\arguments{
\item{x}{A data frame from Analysis Set 1}
}
\value{
A data frame showing the differences between the analysis set and the documentation
}
\description{
Function that checks if the columns in the analysis set match the
columns in the analysis set documentation file
}
