% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_double_columns.R
\name{check_double_columns}
\alias{check_double_columns}
\title{check double columns}
\usage{
check_double_columns(x, y, connector = NULL)
}
\arguments{
\item{x}{Data frame x.}

\item{y}{Data frame y.}

\item{connector}{The connector columns as strings. Also possible as vector.}
}
\value{
Message informing about overlap in columns between the dataframes.
}
\description{
Check whether two dataframes have intersecting column names.
}
\examples{
check_double_columns(mtcars, iris)

}
\seealso{
Other tests: 
\code{\link{duplicates_in_column}()},
\code{\link{test_all_equal}()}
}
\concept{tests}
