% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{check_non_zero_rows}
\alias{check_non_zero_rows}
\title{Check for Non-Zero Rows}
\usage{
check_non_zero_rows(dataframe, collection)
}
\arguments{
\item{dataframe}{A dataframe.}

\item{collection}{A list to store the message if there are 0 rows.}
}
\value{
The updated collection.
}
\description{
This function checks if there are more than 0 rows in the provided dataframe.
If there are 0 rows, a message is added to the provided collection.
}
\examples{
# Create an empty dataframe
dataframe <- data.frame()
collection <- checkmate::makeAssertCollection()
check_non_zero_rows(dataframe, collection)
}
