% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_rows.R
\name{check_rows}
\alias{check_rows}
\title{Check rows}
\usage{
check_rows(df, name = NULL)
}
\arguments{
\item{df}{The data frame whose rows are to be counted}

\item{name}{The name of the data file (this will be printed)}
}
\value{
A message is printed to the console with the number of rows of the data
}
\description{
This function prints the number of rows of a data frame. This function is used
to check that rows are not deleted or doubled unless expected.
}
\examples{
check_rows(mtcars)


}
