% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{check_variable}
\alias{check_variable}
\title{Check variable names}
\usage{
check_variable(name)
}
\arguments{
\item{name}{Variable name to check}
}
\description{
Function to check whether the variable name is according the "stijlgids"
}
