% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{compare_input_output}
\alias{compare_input_output}
\title{Vergelijk input output}
\usage{
compare_input_output(filepath, export_naar_dataframe = FALSE)
}
\arguments{
\item{filepath}{Het complete filepath van het script}

\item{export_naar_dataframe}{Default: FALSE. Whether to export results to a data frame.}
}
\description{
Functie die achterhaald welke bestanden afhankelijk zijn van de output van een script
}
\seealso{
Other Script validatie: 
\code{\link{validate_assertions_present}()},
\code{\link{validate_clear_script_objects}()},
\code{\link{validate_introduction}()},
\code{\link{validate_no_warning_errors}()},
\code{\link{validate_rm_objects}()},
\code{\link{validate_script}()},
\code{\link{validate_style}()},
\code{\link{validate_write_files}()}
}
\concept{Script validatie}
