% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_more_than_1.R
\name{count_more_than_1}
\alias{count_more_than_1}
\title{Count more than 1}
\usage{
count_more_than_1(x)
}
\arguments{
\item{x}{The vector to test}
}
\value{
Number of values greater than 1.
}
\description{
Function to count the number of values greater than 1 in a vector
This function is used in the function Check_columns_for_double_rows
to count duplicate values.
}
\examples{
count_more_than_1(c(1, 1, 4))

}
\concept{vector calculations}
