% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_categorical_details.R
\name{create_categorical_details}
\alias{create_categorical_details}
\title{Create categorical details csv}
\usage{
create_categorical_details(data, mapping)
}
\arguments{
\item{data}{A dataframe for which to create a categorical details csv.}

\item{mapping}{A dataframe containing a mapping named vector, containing preferred fieldnames
Example:
column_names <- c(
mpg = "mpg", cyl = "cyl", disp = "disp", hp = "hp",
drat = "drat", wt = "wt", qsec = "qsec", vs = "vs",
am = "am", gear = "gear", carb = "carb", spare_tire = "spare_tire"}
}
\value{
Dataframe containing categorical details
}
\description{
This function returns a categorical details csv. Containing categorical information about the dataset
}
