% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_summary_table.R
\name{create_dataset_summary_table}
\alias{create_dataset_summary_table}
\title{Create dataset summary statistics table}
\usage{
create_dataset_summary_table(df_input)
}
\arguments{
\item{df_input}{A dataframe for which to create a summary statistics table.}
}
\value{
A tibble with comprehensive summary statistics for each column in the input dataframe.
}
\description{
This function creates a summary statistics table for a dataframe, providing insights into
the nature of the data contained within. It includes detailed statistics for each column,
such as column types, missing value percentages, minimum and maximum values for numeric
columns, patterns for character columns, uniqueness of identifiers, and distributions.
}
