% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_field_info.R
\name{create_field_info}
\alias{create_field_info}
\title{Create field info}
\usage{
create_field_info(
  data,
  raw_data_path = NULL,
  broker = NULL,
  product = NULL,
  public_dataset = NULL
)
}
\arguments{
\item{data}{A dataframe for which to create a field info csv.}

\item{raw_data_path}{A string containing the original location of the original raw file}

\item{broker}{The name of of the organisation or person that distributes the dataset}

\item{product}{The name of the product where this dataset is used in}

\item{public_dataset}{Boolean containing whether the dataset is publicly available
is_primary_key Is_primary_key is variable that can be manually set to TRUE if the dataset contains a primary key.}
}
\value{
Dataframe containing subset info
}
\description{
This function returns a dataframe containing field info information about the dataset
}
