% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_numeric_details.R
\name{create_numeric_details}
\alias{create_numeric_details}
\title{Create numeric details csv}
\usage{
create_numeric_details(data, mapping)
}
\arguments{
\item{data}{A dataframe for which to create a numeric details csv.}

\item{mapping}{A dataframe containing a mapping named vector, containing preferred fieldnames
Example:
column_names <- c(
mpg = "mpg", cyl = "cyl", disp = "disp", hp = "hp",
drat = "drat", wt = "wt", qsec = "qsec", vs = "vs",
am = "am", gear = "gear", carb = "carb", spare_tire = "spare_tire"
)}
}
\value{
Dataframe containing numeric details.
}
\description{
This function returns a numeric details csv. Containing numeric information about the dataset
}
