% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subset_fields.R
\name{create_subset_fields}
\alias{create_subset_fields}
\title{Create subset fields}
\usage{
create_subset_fields(data, mapping)
}
\arguments{
\item{data}{A dataframe for which to create a subsetfields csv.}

\item{mapping}{A dataframe containing a mapping named vector, containing preferred fieldnames
Example:
column_names <- c(
mpg = "mpg", cyl = "cyl", disp = "disp", hp = "hp",
drat = "drat", wt = "wt", qsec = "qsec", vs = "vs",
am = "am", gear = "gear", carb = "carb", spare_tire = "spare_tire"
)}
}
\value{
Dataframe containing subset info
}
\description{
This function returns a subsetfields info df.  Containing subsetfields information about the dataset
}
