% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{determine_author}
\alias{determine_author}
\title{Bepaal Auteur}
\usage{
determine_author(x)
}
\arguments{
\item{x}{Script}
}
\description{
Bepaal auteur van script. Functie voor generate_introduction_validation.
}
