% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{determine_data_files}
\alias{determine_data_files}
\title{Bepaal databestanden}
\usage{
determine_data_files(filepath, testString)
}
\arguments{
\item{filepath}{Het complete filepath van het script}

\item{testString}{"SaveRDS_CSV" of "ReadRDS_CSV" om te checken naar de files}
}
\description{
Functie die achterhaald welke bestanden worden weg gelezen of worden weggeschreven
}
