% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{determine_input_files}
\alias{determine_input_files}
\title{Bepaal inputbestanden}
\usage{
determine_input_files(filepath)
}
\arguments{
\item{filepath}{Het complete filepath van het script}
}
\description{
Functie die achterhaald welke bestanden worden weg gelezen
}
