% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_inputfiles.R
\name{determine_inputfiles}
\alias{determine_inputfiles}
\title{Determine input files}
\usage{
determine_inputfiles(path)
}
\arguments{
\item{path}{Het complete path van het script}
}
\description{
Determine which files are read in.
}
