% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{determine_output_files}
\alias{determine_output_files}
\title{Bepaal outputbestanden}
\usage{
determine_output_files(filepath)
}
\arguments{
\item{filepath}{Het complete filepath van het script}
}
\description{
Functie die achterhaald welke bestanden worden weg geschreven
}
