% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_compare.R
\name{df_compare}
\alias{df_compare}
\title{df compare}
\usage{
df_compare(df1, df2)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}
}
\value{
A dataframe with comparisons between the input dataframes..
}
\description{
Compare two dataframes in order to check whether their properties are identical.
}
