% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_inspect_without_warnings.R
\name{df_inspect_without_warnings}
\alias{df_inspect_without_warnings}
\title{df_inspect_without_warnings}
\usage{
df_inspect_without_warnings(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
suppressWarnings(df_inspect(df, print_results = F))
}
\description{
Wrapper for df_inspect to suppress warnings.
}
\examples{
\dontrun{df_inspect_without_warnings(df)}
}
