% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drop_NA_column_names.R
\name{drop_na_column_names}
\alias{drop_na_column_names}
\title{Drop NA column names}
\usage{
drop_na_column_names(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
dataframe without columns that are NA
}
\description{
Deletes columns whose name is NA or whose name is empty
}
