% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_in_column.R
\name{duplicates_in_column}
\alias{duplicates_in_column}
\title{Duplicates in column}
\usage{
duplicates_in_column(df, col)
}
\arguments{
\item{df}{Data frame.}

\item{col}{Column name.}
}
\value{
Rows containing duplicated values.
}
\description{
Searches for duplicates in a data frame column.
}
\examples{
duplicates_in_column(mtcars, "mpg")

}
\seealso{
Other tests: 
\code{\link{check_double_columns}()},
\code{\link{test_all_equal}()}
}
\concept{tests}
