% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_filepaths_readrds_csv.R
\name{extract_filepaths_readrds_csv}
\alias{extract_filepaths_readrds_csv}
\title{Extract filepaths readrds_csv}
\usage{
extract_filepaths_readrds_csv(bestandspad)
}
\arguments{
\item{bestandspad}{Het complete bestandspad van het script}
}
\description{
Functie die voor Genereer_outputbestanden_database bestandspaden achterhaald van ingelezen bestanden
}
