% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{extract_output_file}
\alias{extract_output_file}
\title{Extraheer output filepath}
\usage{
extract_output_file(regel, testString)
}
\arguments{
\item{regel}{SaveRDS_CSV of ReadRDS_CSV regel die moet worden omgeschreven naar filepath}

\item{testString}{"SaveRDS_CSV" of "ReadRDS_CSV" om te vergelijken of het een Read of Save file is}
}
\description{
Functie die SaveRDS_CSV omschrijft naar kloppend filepath
}
