% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{generate_introduction_validation}
\alias{generate_introduction_validation}
\title{Genereer introductie validatie}
\usage{
generate_introduction_validation(Scripts, Rapportage = TRUE)
}
\arguments{
\item{Scripts}{List met filepath(en) van de/het script(s) die gevalideerd moeten worden}

\item{Rapportage}{Mogelijkheid om samenvattende rapportage te printen bij het uitvoeren van het script}
}
\description{
Functie om scripts te testen op de volledigheid van de introductie. De functie geeft
een dataframe terug met daarin de resultaten van de verschillende tests worden
uitgevoerd. Optioneel kan er een kleine rapportage worden geprint. Deze functie wordt gebruikt
in de validate_script functie en in 05. Rapporten/XX Rapport SA Controle introducties en bestandsnamen.R
}
