% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{generate_output_files_database}
\alias{generate_output_files_database}
\title{Genereer outputbestanden database}
\usage{
generate_output_files_database()
}
\description{
Functie die voor de stappen 1,2,3 en 5 de R bestanden die in script in leest
opslaat in een database. Deze database kan gevonden worden op de volgende locatie
Network_directory/XX. Testbestanden/Overzicht ingelezen bestanden.csv
}
