% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_summary_table.R
\name{get_first_element_class}
\alias{get_first_element_class}
\title{Retrieve the class of the first element of a vector}
\usage{
get_first_element_class(input_vector)
}
\arguments{
\item{input_vector}{A vector whose first element's class is to be retrieved.}
}
\value{
The class of the first element of the input vector.
}
\description{
Retrieve the class of the first element of a vector
}
\examples{
# Get the class of the first element in a numeric vector
get_first_element_class(c(1, 2, 3)) # Returns "numeric"

# Get the class of the first element in a character vector
get_first_element_class(c("apple", "banana", "cherry")) # Returns "character"
}
