% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Get values of column}
\usage{
get_values(analyseset, column)
}
\arguments{
\item{analyseset}{The analysisset}

\item{column}{Column to get values from.}
}
\description{
A function to determine what kind of values are present in columns.
}
