% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{global_xml_parsed_content}
\alias{global_xml_parsed_content}
\title{Retrieve the XML parsed content from an XML document}
\usage{
global_xml_parsed_content(source_file)
}
\arguments{
\item{source_file}{XML document}
}
\description{
Function to retrieve the XML parsed content from an XML document
}
