% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_globale_identifiers.R
\name{has_globale_identifiers}
\alias{has_globale_identifiers}
\title{Has globale identifiers}
\usage{
has_globale_identifiers(df)
}
\arguments{
\item{df}{Data frame.}
}
\description{
Check whether dataframe has "globale identifiers".
In the scope of the vu student analytics project, these are:
INS_Studentnummer, INS_Opleidingsnaam_2002, INS_Opleidingscode_actueel,
INS_Inschrijvingsjaar, INS_Inschrijvingsjaar_EOI.
}
