% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_join_pairs.R
\name{identify_join_pairs}
\alias{identify_join_pairs}
\title{Identify Possible Join Pairs Between Data Frames}
\usage{
identify_join_pairs(..., similarity_cutoff = 0.2)
}
\arguments{
\item{...}{A list of two data frames.}

\item{similarity_cutoff}{The minimal percentage of overlap between the distinct values in the columns.}
}
\value{
A data frame showing candidate join pairs.
}
\description{
This function identifies potential join pairs between two data frames based on the overlap
between the distinct values in their columns. It returns a data frame showing the possible join pairs.
}
\examples{
identify_join_pairs(iris, iris3)

}
