% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{make_lint_object}
\alias{make_lint_object}
\title{Make lint object}
\usage{
make_lint_object(expr, source_file, type)
}
\arguments{
\item{expr}{XML parsed content}

\item{source_file}{XML document}

\item{type}{Lint type}
}
\description{
Function to make lint objects for the Validate_variable_name
}
