% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_important_in_project.R
\name{package_important_in_project}
\alias{package_important_in_project}
\title{package_important_in_project}
\usage{
package_important_in_project(
  package = "vusa",
  type = 2,
  tellen = "aantal_keer",
  diepte = 1,
  kolommen = c("00. Downloaden", "01. Inlezen", "02. Manipuleren",
    "03. Analyseset maken", "04. Analyseren", "05. Rapporten")
)
}
\arguments{
\item{package}{De package waarvan je wilt uitzoeken hoe vaak die wordt gebruikt}

\item{type}{type functies: 2 voor functies beschikbaar met '::', 3 voor functies
beschikbaar met ':::' en -1 voor het verschil er tussen}

\item{tellen}{'aantal_keer' voor hoe vaak de functie voor komt, 'aantal_bestanden'
voor in hoe veel bestanden een functie voorkomt: Als een functie dan meerdere
keren in een bestand voorkomt wordt er dus maar een keer geteld}

\item{diepte}{Default = 1. Geeft aan hoe veel submappen er in de uiteindlijke
dataframe komen. De range is van 1-3. Bij 3 Komen dus bvb kolommen als :
"Map1/Map2/Map3". Voor verdere inspectie is het beter deze gehele functie in
die submap te runnen}

\item{kolommen}{which directories to check.}
}
\value{
Dataframe met de functies van de package als rijen, en submappen van
huidig project als kolommen. De waardes in de dataframe geven aan hoe vaak een
functie in welke map gebruikt wordt
}
\description{
Een functie die gerund kan worden in een project (folder). Bij
selectie van een package (default: "vusa") wordt er een dataframe gereturnt
waarin elke functie uit de package een row is. Elke map in het project (folder)
is een kolom. De waardes in het dataframe geven aan hoe vaak een functie in welke
map voor komt. Zo kan worden onderzocht welke functies waar in een project
gebruikt worden.
-De functie checkt automatisch alle submappen van een project, zorg dat je
deze dus altijd vanuit de hoofd-folder runt.
-Omdat de functie nog niet geoptimaliseerd is (hij checkt nu alle scripts opnieuw
voor elke functie) duurt het even voordat alle scripts zijn doorlopen.
200 functies duren +/- 10 minuten.
-Als je meer informatie over een specifieke submap wil kun je de functie dus
nog een keer in die submap specifiek laten runnen
}
\examples{
\dontrun{df_functies_in_package <- package_important_in_project()}
}
