% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_last_used.R
\name{print_last_used}
\alias{print_last_used}
\title{print_last_used}
\usage{
print_last_used(path = NULL)
}
\arguments{
\item{path}{The filepath}
}
\description{
Prints the data on which the given script was last run succesfully.
}
