% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{rapport_result_test_introduction}
\alias{rapport_result_test_introduction}
\title{Rapporteer Uitkomst test introductie}
\usage{
rapport_result_test_introduction(obj, melding)
}
\arguments{
\item{obj}{Object dat moet worden gerapporteerd}

\item{melding}{Melding die moet worden weergeven}
}
\description{
Functie om de uitkomst van een test te melden. Functie voor generate_introduction_validation.
}
