% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_df_columns.R
\name{ratio_df_columns}
\alias{ratio_df_columns}
\title{relatief_belangrijke_waardes_columns}
\usage{
ratio_df_columns(
  df,
  columns = c("00. Downloaden", "01. Inlezen", "02. Manipuleren", "03. Analyseset maken",
    "04. Analyseren", "05. Rapporten")
)
}
\arguments{
\item{df}{Dataframe}

\item{columns}{De columns uit het dataframe waarover de relatieve belangrijkheid
wordt berekent (en waar tevens voor elke column een extra column wordt toegevoegd)}
}
\value{
Dataframe met toegevoegde columns. In deze columns staat een waarde
tussen 0.0 en 1.0.
1.0 = 100\% van de keren dat deze waarde voor komt is binnen deze column
0.0 =  0\% van de keren dat deze waarde voor komt is binnen deze column
}
\description{
Een ondersteunende functie voor package_belangrijk_in_project,
maar kan voor elke dataframe gebruikt worden. Met gebruik van deze functie
worden columns toegevoegd die aangeven hoe belangrijk een (tel) waarde is
binnen een selectie columns. De functie werkt als volgt:
\enumerate{
\item Specificeer de columns waar het om gaat als 'columns'
---De functie maakt de column 'sum' aan met de opsomming van alle columns
---Voor elke column uit 'columns' wordt een column aangemaakt met de originele
waarde uit de column gedeeld door de opsomming (column/sum)
=> Dit getal geeft aan hoe uniek een waarde uit een row is:
1.0 = 100\% van de keren dat deze waarde voor komt is binnen deze column
0.0 =  0\% van de keren dat deze waarde voor komt is binnen deze column
}
}
\examples{
\dontrun{dfTest <- ratio_df_columns(dfTest, c("column1", "column2"))}
}
