% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_content_parameter.R
\name{regex_content_parameter}
\alias{regex_content_parameter}
\title{Construct Regex for Matching Function Parameter Content}
\usage{
regex_content_parameter(parameter)
}
\arguments{
\item{parameter}{The parameter whose value is to be searched in a function.}
}
\value{
A regex pattern as a character string.
}
\description{
This function constructs a regex pattern for matching the content of a parameter in a function.
It uses the \code{base::paste0} function to construct the regex pattern.
}
\examples{
# Create a parameter name
parameter <- "my_parameter"
# Construct a regex pattern for matching the content of the parameter
pattern <- regex_content_parameter(parameter)

}
