% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_time.R
\name{regex_time}
\alias{regex_time}
\title{Generate regular expression of a time.}
\usage{
regex_time(format = "hh:mm")
}
\arguments{
\item{format}{The format of the time. Possible values are:
\itemize{
\item{"hh:mm"}: to generate "09:05".
\item{"h:m"}: to generate "9:5".
\item{"hh:mm:ss"}: to generate "09:05:00".
\item{"h:m:s"}: to generate "9:5:0".
\item{"hh:mm:ss AM/PM"}: to generate "09:05:00 AM".
\item{"h:m:s AM/PM"}: to generate "9:5:0 AM".
}}
}
\value{
A regular expression.
}
\description{
This function generates a regular expression for time based on the input format.
}
\examples{
regex_time("hh:mm")
regex_time("h:m")
regex_time("hh:mm:ss")
regex_time("h:m:s")
regex_time("hh:mm:ss AM/PM")
regex_time("h:m:s AM/PM")

}
