% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_function_calls.R
\name{retrieve_function_calls}
\alias{retrieve_function_calls}
\title{retrieve_function_calls}
\usage{
retrieve_function_calls(script_name)
}
\arguments{
\item{script_name}{The script to search functions in}
}
\value{
dataframe
}
\description{
retrieve_function_calls
}
