% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_functions_and_packages.R
\name{retrieve_functions_and_packages}
\alias{retrieve_functions_and_packages}
\title{Retrieve functions and packages}
\usage{
retrieve_functions_and_packages(path)
}
\arguments{
\item{path}{The complete path of the script.}
}
\value{
Used_functions
}
\description{
Retrieves functions and their corresponding packages used in a given script.
}
