% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_package_usage.R
\name{retrieve_package_usage}
\alias{retrieve_package_usage}
\title{Retrieve packages that are loaded in a script}
\usage{
retrieve_package_usage(script_name)
}
\arguments{
\item{script_name}{The path to the R script}
}
\value{
dataframe
}
\description{
Retrieve packages that are loaded in a script
}
