% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_sourced_scripts.R
\name{retrieve_sourced_scripts}
\alias{retrieve_sourced_scripts}
\title{retrieve_sourced_scripts}
\usage{
retrieve_sourced_scripts(script_name)
}
\arguments{
\item{script_name}{The main script to search}
}
\value{
dataframe
}
\description{
retrieve_sourced_scripts
}
