% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_string_assignments.R
\name{retrieve_string_assignments}
\alias{retrieve_string_assignments}
\title{retrieve_string_assignments}
\usage{
retrieve_string_assignments(script_name)
}
\arguments{
\item{script_name}{The script to search objects in}
}
\value{
dataframe
}
\description{
retrieve_string_assignments
}
