% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_mtcars_testfile.R
\name{return_mtcars_testfile}
\alias{return_mtcars_testfile}
\title{Read and return the mtcars testfile}
\usage{
return_mtcars_testfile()
}
\value{
returns mtcars_test dataframe
}
\description{
Gets the modified rds dataset for testing assertions.
}
