% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{strip_names}
\alias{strip_names}
\title{Strip punctuation characters from variable name}
\usage{
strip_names(x)
}
\arguments{
\item{x}{Variable name to strip}
}
\description{
Function to strip the punctuation characters
}
