% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_all_equal.R
\name{test_all_equal}
\alias{test_all_equal}
\title{Test all equal}
\usage{
test_all_equal(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector to test.}

\item{na.rm}{default: FALSE. exclude NAs from the test.}
}
\value{
Boolean result of the test
}
\description{
Test whether all values in a vector are equal.
}
\examples{
test_all_equal(c(5, 5, 5))

test_all_equal(c(5, 6, 3))

}
\seealso{
Other tests: 
\code{\link{check_double_columns}()},
\code{\link{duplicates_in_column}()}
}
\concept{tests}
