% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_csv.R
\name{test_csv}
\alias{test_csv}
\title{Test csv}
\usage{
test_csv(dataloc = "")
}
\arguments{
\item{dataloc}{De locatie waar het bestand naar weggeschreven wordt}
}
\value{
Een dataframe dat de verschillen tussen de kolomnamen en classes van de twee
inleesmethode weergeeft
}
\description{
Vergelijkt de kolomnamen tussen read_csv2 en read.csv2. Deze functie is gemaakt om
te inventariseren wat de impact is van het gebruiken van de readr package om csv's in
te lezen.
}
