% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{test_file_on_documentation}
\alias{test_file_on_documentation}
\title{Test bestand op documentatie}
\usage{
test_file_on_documentation(
  filepath,
  export_naar_dataframe = FALSE,
  processbar = NULL
)
}
\arguments{
\item{filepath}{Het verkorte filepath van het data- of documentatie bestand}

\item{export_naar_dataframe}{Default: FALSE. Whether to export results to a data frame.}

\item{processbar}{Mogelijkheid om processbar te weergeven bij gebruik van map en deze functie.}
}
\description{
Functie om per document het documentatiebestand in te lezen, de data in te lezen
en de assertion uit te voeren. Zowel het databestand(.rds) als het documentatiebestand(.csv)
kan worden meegegeven als filepath. Deze functie wordt zowel gebruikt in de validate_script
functie als in XX. Testraamwerk/Testen_scripts.R.
}
