% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_id.R
\name{unique_id}
\alias{unique_id}
\title{unique id}
\usage{
unique_id(x, ...)
}
\arguments{
\item{x}{vector or dataframe.}

\item{...}{optional variables, e.g. name of column or a vector of names.}
}
\value{
Boolean whether variable is a unique identifier.
}
\description{
Check if parsed variable is a unique identifier.
This function was adapted from:
Source: https://edwinth.github.io/blog/unique_id/
}
\examples{
unique_id(iris, Species)

mtcars$name <- rownames(mtcars)
unique_id(mtcars, name)

}
