% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pattern_r.R
\name{find_pattern_r}
\alias{find_pattern_r}
\title{Find pattern in R scripts}
\usage{
find_pattern_r(pattern, path = ".", case.sensitive = TRUE, comments = FALSE)
}
\arguments{
\item{pattern}{Pattern to search on}

\item{path}{de folder waar gezocht moet worden, default working directory}

\item{case.sensitive}{Whether pattern is case sensitve or ntot}

\item{comments}{whether to search in commented lines}
}
\value{
dataframe met R script paths
}
\description{
Function to search for a pattern in R scripts.
}
\examples{
\dontrun{
find_pattern_r(pattern = "DUO_", path = "05. Rapporten/")
find_pattern_r("DUO_", c("05. Rapporten/", "02. Manipuleren/"))
}
}
