% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Get values of column}
\usage{
get_values(df, column)
}
\arguments{
\item{df}{The dataframe}

\item{column}{Column to get values from.}
}
\value{
The class of the column values
}
\description{
A function to determine what kind of values are present in columns.
}
\examples{
get_values(mtcars, "mpg")


}
