% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_complete_cases.R
\name{md_complete_cases}
\alias{md_complete_cases}
\title{MD complete cases}
\usage{
md_complete_cases(data, digits = 1)
}
\arguments{
\item{data}{The data frame.}

\item{digits}{Default: 1. number of digits for rounding.}
}
\value{
Message with the number of rows, number of rows with missing values and the percentage of complete rows.
}
\description{
Print the complete cases of the data.
}
\examples{
# example code
md_complete_cases(iris)

iris$Sepal.Length[5] <- NA_character_
md_complete_cases(iris)

}
