% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{regex_content_parameter}
\alias{regex_content_parameter}
\title{Regex content parameter}
\usage{
regex_content_parameter(parameter)
}
\arguments{
\item{parameter}{De parameter waarvan de waarde gezocht moet worden in een functie}
}
\description{
Functie die een regex construeert voor het matchen van de content uit een
parameter van een functie
}
