% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect_in_file.R
\name{str_detect_in_file}
\alias{str_detect_in_file}
\title{Detect string in file}
\usage{
str_detect_in_file(file, pattern, only_comments = FALSE, collapse = FALSE)
}
\arguments{
\item{file}{Path to file.}

\item{pattern}{Pattern to match.}

\item{only_comments}{default FALSE. Whether to only search in commented lines.}

\item{collapse}{default: FALSE: search file line by line.
If true, then pattern is search in the entire file at once after collapsing.
(only_comments does not work when collapse is set to TRUE)}
}
\value{
Boolean whether pattern exists in file.
}
\description{
Detect string in file
}
